+++ titan/parse.py
@@ -575,83 +575,6 @@ def parse_function_name(header: str):
     return prefix
 
 
-def parse_identifier(identifier: str, is_db_scoped=False) -> FQN:
-    # TODO: This needs to support periods and question marks in double quoted identifiers
-    scoped_name, param_str = identifier.split("?") if "?" in identifier else (identifier, "")
-    params = {}
-    if param_str:
-        for param in param_str.split("&"):
-            k, v = param.split("=")
-            params[k] = v
-
-    arg_types = []
-    if "(" in scoped_name:
-        args_start = scoped_name.find("(")
-        scoped_name, args_str = scoped_name[:args_start], scoped_name[args_start:]
-        arg_types = [arg.strip() for arg in args_str.strip("()").split(",")]
-
-    try:
-        name_parts = list(FullyQualifiedIdentifier.parse_string(scoped_name, parse_all=True))
-    except pp.ParseException:
-        raise pp.ParseException(f"Failed to parse identifier: {identifier}")
-    if len(name_parts) == 1:
-        return FQN(
-            name=name_parts[0],
-            params=params,
-            arg_types=arg_types,
-        )
-    elif len(name_parts) == 2:
-        if is_db_scoped:
-            return FQN(
-                database=name_parts[0],
-                name=name_parts[1],
-                params=params,
-                arg_types=arg_types,
-            )
-        else:
-            return FQN(
-                schema=name_parts[0],
-                name=name_parts[1],
-                params=params,
-                arg_types=arg_types,
-            )
-    elif len(name_parts) == 3:
-        return FQN(
-            database=name_parts[0],
-            schema=name_parts[1],
-            name=name_parts[2],
-            params=params,
-            arg_types=arg_types,
-        )
-    elif len(name_parts) == 4:
-        params["entity"] = name_parts[3]
-        return FQN(
-            database=name_parts[0],
-            schema=name_parts[1],
-            name=name_parts[2],
-            params=params,
-            arg_types=arg_types,
-        )
-    raise Exception(f"Failed to parse identifier: {identifier}")
-
-
-# NOTE: can't put this into identifiers.py:URN because of circular import
-def parse_URN(urn_str: str) -> URN:
-    parts = urn_str.split(":")
-    if len(parts) != 4:
-        raise Exception(f"Invalid URN string: {urn_str}")
-    if parts[0] != "urn":
-        raise Exception(f"Invalid URN string: {urn_str}")
-    resource_label, fqn_str = parts[3].split("/", 1)
-    resource_type = resource_type_for_label(resource_label)
-    fqn = parse_identifier(fqn_str, is_db_scoped=(resource_label == "schema"))
-    return URN(
-        account_locator=parts[2],
-        resource_type=resource_type,
-        fqn=fqn,
-    )
-
-
 def _parse_copy_into(sql: str):
     """
     /* Standard data load */
