+++ tools/reset_test_account.py
@@ -22,7 +22,7 @@ def load_and_run_config(conn, file, run_mode, allowlist):
         run_mode=run_mode,
         allowlist=allowlist,
         resources=collect_resources_from_config(config),
-        dry_run=True,
+        # dry_run=True,
     )
     plan = bp.plan(conn)
     print_plan(plan)
